requirements.txt
langchain
langchain_core
langchain-groq
langchain-community
python-dotenv
streamlit
